first functional version using typescript
Only bar.js is bar.ts -- very simple port
