+++ package.json
@@ -30,9 +30,13 @@
   },
   "devDependencies": {
     "@babel/core": "^7.5.5",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-proposal-export-namespace-from": "^7.5.2",
     "@babel/plugin-transform-object-assign": "^7.2.0",
     "@babel/preset-env": "^7.5.5",
+    "@types/jquery": "^3.3.31",
+    "@typescript-eslint/eslint-plugin": "^2.3.0",
+    "@typescript-eslint/parser": "^2.3.0",
     "ajv": "^6.10.2",
     "babel-loader": "^8.0.6",
     "core-js": "^3.2.1",
@@ -56,6 +60,8 @@
     "npm-check-updates": "^3.1.21",
     "qunit": "^2.9.2",
     "style-loader": "^0.23.1",
+    "ts-loader": "^6.1.2",
+    "typescript": "^3.6.3",
     "webpack": "^4.39.3"
   },
   "directories": {
